decoder: narrow the motion grids to (i16,i16) - 2.90 GB of pool clear removed
The largest remaining item in docs/copy-inventory.md, deferred twice and
priced at 276 MB per 60-frame decode. Landing it required answering the
objection that deferred it, which was NOT about the width:

  ref_idx is COMPARED. A motion vector is COMPUTED WITH -- median
  prediction, mvd addition, temporal-direct scaling. Narrowing the grid
  pulls i16 into arithmetic that was i32, and since overflow-checks is
  off in release, an intermediate that overflows now wraps silently.

The answer is that Rust has no implicit numeric coercion, so a narrow
grid makes the COMPILER the proof: every read must widen explicitly or
it does not build. That is exactly what happened -- it rejected the
temporal-direct scaling at mb16.rs:6777, `dsf * mvc.0`, which is the
precise site the deferral was worried about. Storage is narrow; all
arithmetic still runs at i32; only the final store narrows, and
`pack_mb` already narrowed the same values unconditionally for the
deblock path, so the store-side risk was already taken.

  decoder mv_y, mv1            (i32,i32) -> (i16,i16)
  encoder mv_y, mv1_y, MbState.mv_y, RefFrame.mv/.mv1
  BlockInfo.mv/.mv1            &[(i32,i32)] -> &[(i16,i16)]
  gmv job arrays + locals      [(i32,i32);16] -> [(i16,i16);16]
  MvField (public diagnostic)  kept i32, widened at capture

Measured on _dprof/shields__main.264 (1260 frames, 720p Main), where
every bucket cross-checks against the geometry (57,600 4x4 blocks x 4 B
= 225.0 KB/call; 3,600 MBs x 32 B = 112.5 KB/call):

  per-picture pool clear     11,567,052,000 -> 8,664,012,000 B  (-25.1%)
  per-reference motion clone  3,483,648,000 -> 1,935,360,000 B  (-44.4%)

Also fixes the REF_BUILD counter, which hardcoded size_of::<(i32,i32)>()
and would have reported double the moment the grids narrowed. It now
reads size_of_val on the slice, so the compiler supplies the width and
the instrument cannot drift from the code it prices.

Gates: 68/68 byte-identical vs ffmpeg, 29 test suites pass, workspace
--all-targets builds warning-free.

+++ crates/rusty_h264-common/examples/deblock_anatomy.rs
@@ -28,7 +28,7 @@ use rusty_h264_common::deblock::{filter_frame, BlockInfo};
 struct Grid {
     inter: Vec<bool>,
     nnz: Vec<u8>,
-    mv: Vec<(i32, i32)>,
+    mv: Vec<(i16, i16)>,
     ref_id: Vec<i8>,
     t8x8: Vec<bool>,
 }
@@ -42,7 +42,7 @@ fn grid(kind: &str, mb_w: usize, mb_h: usize) -> Grid {
     let mut g = Grid {
         inter: vec![true; n],
         nnz: vec![0u8; n],
-        mv: vec![(0i32, 0i32); n],
+        mv: vec![(0i16, 0i16); n],
         ref_id: vec![0i8; n],
         t8x8: Vec::new(),
         };
@@ -67,7 +67,7 @@ fn grid(kind: &str, mb_w: usize, mb_h: usize) -> Grid {
         "inter-bs0" => {
             for (i, m) in g.mv.iter_mut().enumerate() {
                 // ±1 quarter-pel: |Δ| < 4 for every neighbour pair ⇒ bS 0.
-                *m = ((i % 2) as i32, ((i / 2) % 2) as i32);
+                *m = ((i % 2) as i16, ((i / 2) % 2) as i16);
             }
         }
         // Real content, unlike every scenario above, mixes intra/inter, coded and
@@ -87,7 +87,7 @@ fn grid(kind: &str, mb_w: usize, mb_h: usize) -> Grid {
                 let r = rnd();
                 g.inter[i] = r & 7 != 0; // ~1 in 8 blocks intra
                 g.nnz[i] = if r & 0x30 != 0 { (r >> 8 & 7) as u8 } else { 0 };
-                g.mv[i] = (((r >> 11) & 15) as i32 - 8, ((r >> 15) & 15) as i32 - 8);
+                g.mv[i] = (((r >> 11) & 15) as i16 - 8, ((r >> 15) & 15) as i16 - 8);
                 g.ref_id[i] = if g.inter[i] { ((r >> 19) & 1) as i8 } else { i8::MIN };
             }
         }
